[U] Update method call

+++ src/main/java/org/hydev/veracross/analyzer/api/nodes/veracross/courseinfo/NodeCourseInfoGetRating.java
@@ -11,6 +11,7 @@ import org.hydev.veracross.sdk.VeracrossHttpClient;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static java.lang.Integer.parseInt;
 import static java.lang.Long.parseLong;
 import static org.hydev.veracross.analyzer.VAConstants.LENGTH_TOKEN;
 import static org.hydev.veracross.analyzer.database.model.CourseInfoRating.getByCourse;
@@ -70,7 +71,7 @@ public class NodeCourseInfoGetRating extends JsonApiNode<NodeCourseInfoGetRating
             switch (data.condition)
             {
                 case "user": ratings = getByUser(parseLong(data.value)); break;
-                case "course": ratings = getByCourse(parseLong(data.value)); break;
+                case "course": ratings = getByCourse(parseInt(data.value)); break;
                 default: return new JsonKnownError("What?");
             }
